Unit stats are calculated from gear
Unit damage and armor are calculated off gear.
Unit now grabs its own UnitInventory.
Inventory now has text to indicate armor/damage

+++ Assets/Scripts/PlayerController.cs
@@ -85,7 +85,7 @@ public class PlayerController : UnitController
                 if (hit.collider.gameObject.tag == "Enemy" && Input.GetMouseButtonDown(0) && GetCurrentHP() > 0)
                 {
                     EnemyController enemy = hit.collider.gameObject.GetComponent<EnemyController>();
-                    enemy.TakeDamage(unit.damage);
+                    enemy.TakeDamage((int)unit.damage);
                     gameController.UpdateEventLog("Player dealt " + unit.damage + " damage to " + enemy.name);
                     gameController.SetPlayerTurn(false);
                 }
